Application is now functional

+++ src/main/java/com/alexpages/ebankingapi/model/transaction/Transaction.java
@@ -18,7 +18,7 @@ public class Transaction {
     private String id;                      // Unique identifier (e.g. 89d3o179-abcd-465b-o9ee-e2d5f6ofEld46)
     @Enumerated(EnumType.STRING)
     private Currency currency;              // currency (eg GBP 100-, CHF 75)
-    private BigDecimal amount;              // Amount with currency (eg GBP 100-, CHF 75)
+    private double amount;                  // Amount with currency (eg GBP 100-, CHF 75)
     private String iban;                    // Account IBAN (eg. CH93-0000-0000-0000-0000-0)
     @JsonFormat(shape = JsonFormat.Shape.STRING,
             pattern = "dd-MM-yyyy")
